jean: Ajout de la version 4 implementant l'algorithme de Pagerank avec la methode de Gauss Seidel descendant. Modification du README et du Makefile en consequence

+++ src/algo_PR_v1.c
@@ -240,11 +240,11 @@ int main(int argc, char const *argv[]) {
 	printf("\n*******************************************\n");
 	printf("\nNombre d'iterations pour la convergeance : %d\n", nbr_iterations_convergence);
 	printf("\nTemps de lecture du fichier en CPU ticks : %lu\n", fin_lecture_t - debut_lecture_t);
-	printf("Temps de lecture du fichier en ms : %lu\n", (fin_lecture_t - debut_lecture_t)*1000/CLOCKS_PER_SEC);
+	printf("Temps de lecture du fichier en ms : %lf\n", (double)(fin_lecture_t - debut_lecture_t)*1000/CLOCKS_PER_SEC);
 	printf("Temps de calcul du pagerank en CPU ticks : %lu\n", fin_calcul_t - debut_calcul_t);
-	printf("Temps de calcul du pagerank en ms : %lu\n", (fin_calcul_t - debut_calcul_t)*1000/CLOCKS_PER_SEC);
+	printf("Temps de calcul du pagerank en ms : %lf\n", (double)(fin_calcul_t - debut_calcul_t)*1000/CLOCKS_PER_SEC);
 	printf("\nTemps total de calcul du pagerank en CPU ticks : %lu\n", fin_calcul_t - debut_lecture_t);
-	printf("Temps total de calcul du pagerank en ms : %lu\n", (fin_calcul_t - debut_lecture_t)*1000/CLOCKS_PER_SEC);
+	printf("Temps total de calcul du pagerank en ms : %lf\n", (double)(fin_calcul_t - debut_lecture_t)*1000/CLOCKS_PER_SEC);
 	printf("\nQuantité de memoire allouée dynamiquement en octets : %d\n", quantite_memoire_allouee);
 	printf("Quantité de memoire libèrée dynamiquement en octets : %d\n", quantite_memoire_liberee);
